refactor: move RemoteConnectionService interface out of root package
