+++ remote_connection.go
@@ -1,39 +1,8 @@
 package influxdb
 
 import (
-	"context"
-
 	"github.com/influxdata/influxdb/v2/kit/platform"
 )
-
-type RemoteConnectionService interface {
-	// ListRemoteConnections returns all info about registered remote InfluxDB connections matching a filter.
-	ListRemoteConnections(context.Context, RemoteConnectionListFilter) (*RemoteConnections, error)
-
-	// CreateRemoteConnection registers a new remote InfluxDB connection.
-	CreateRemoteConnection(context.Context, CreateRemoteConnectionRequest) (*RemoteConnection, error)
-
-	// ValidateNewRemoteConnection validates that the given settings for a remote InfluxDB connection are usable,
-	// without persisting the connection info.
-	ValidateNewRemoteConnection(context.Context, CreateRemoteConnectionRequest) error
-
-	// GetRemoteConnection returns metadata about the remote InfluxDB connection with the given ID.
-	GetRemoteConnection(context.Context, platform.ID) (*RemoteConnection, error)
-
-	// UpdateRemoteConnection updates the settings for the remote InfluxDB connection with the given ID.
-	UpdateRemoteConnection(context.Context, platform.ID, UpdateRemoteConnectionRequest) (*RemoteConnection, error)
-
-	// ValidateUpdatedRemoteConnection validates that a remote InfluxDB connection is still usable after applying the
-	// given update, without persisting the new info.
-	ValidateUpdatedRemoteConnection(context.Context, platform.ID, UpdateRemoteConnectionRequest) error
-
-	// DeleteRemoteConnection deletes all info for the remote InfluxDB connection with the given ID.
-	DeleteRemoteConnection(context.Context, platform.ID) error
-
-	// ValidateRemoteConnection checks that the remote InfluxDB connection with the given ID is still usable
-	// with its persisted settings.
-	ValidateRemoteConnection(context.Context, platform.ID) error
-}
 
 // RemoteConnection contains all info about a remote InfluxDB instance that should be returned to users.
 // Note that the auth token used by the request is *not* included here.
