+++ remotes/transport/http_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/influxdata/influxdb/v2"
 	"github.com/influxdata/influxdb/v2/kit/feature"
 	"github.com/influxdata/influxdb/v2/kit/platform"
-	"github.com/influxdata/influxdb/v2/mock"
+	"github.com/influxdata/influxdb/v2/remotes/mock"
 	"github.com/stretchr/testify/assert"
 	tmock "github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -230,7 +230,7 @@ func TestRemoteConnectionHandler(t *testing.T) {
 func newTestServer(t *testing.T) (*httptest.Server, *mock.MockRemoteConnectionService) {
 	ctrlr := gomock.NewController(t)
 	svc := mock.NewMockRemoteConnectionService(ctrlr)
-	server := annotatedTestServer(NewRemoteConnectionHandler(zaptest.NewLogger(t), svc))
+	server := annotatedTestServer(newRemoteConnectionHandler(zaptest.NewLogger(t), svc))
 	return httptest.NewServer(server), svc
 }
